Connect to FatsecretAPI

+++ app/src/main/java/com/example/ctracker/ApiService/FoodApiService.kt
@@ -6,7 +6,7 @@ import retrofit2.http.*
 data class FoodResponse(
     val name: String,
     val food_id: Int,
-    val calorie: Int,
+    val calorie: Float,
     val carbs: Float,
     val fats: Float,
     val protein: Float
