+++ app/src/main/java/com/example/ctracker/entity/Food.kt
@@ -3,7 +3,7 @@ package com.example.ctracker.entity
 data class Food(
     val id: Int,
     val name: String,
-    val calories: Int,
+    val calories: Float,
     val fat: Float,
     val carb: Float,
     val protein: Float
