+++ app/src/main/java/com/example/ctracker/viewmodel/AddItemViewModel.kt
@@ -40,7 +40,7 @@ class AddItemViewModel(private val index: Int, private val mealType: Int) : View
                     food.value = Food(
                         id = foodResponse.food_id,
                         name = foodResponse.name,
-                        calories = foodResponse.calorie,
+                        calories = foodResponse.calorie.toFloat(),
                         protein = foodResponse.protein,
                         fat = foodResponse.fats,
                         carb = foodResponse.carbs
